+++ keewe-api/src/main/java/ccc/keeweapi/service/user/ProfileApiService.java
@@ -3,7 +3,20 @@ package ccc.keeweapi.service.user;
 import static ccc.keewecore.consts.KeeweConsts.MY_PAGE_TITLE_LIMIT;
 
 import ccc.keeweapi.component.ProfileAssembler;
-import ccc.keeweapi.dto.user.*;
+import ccc.keeweapi.dto.user.AllAchievedTitleResponse;
+import ccc.keeweapi.dto.user.BlockUserResponse;
+import ccc.keeweapi.dto.user.FollowToggleResponse;
+import ccc.keeweapi.dto.user.FollowUserListResponse;
+import ccc.keeweapi.dto.user.InterestsResponse;
+import ccc.keeweapi.dto.user.MyBlockUserListResponse;
+import ccc.keeweapi.dto.user.MyPageTitleResponse;
+import ccc.keeweapi.dto.user.OnboardRequest;
+import ccc.keeweapi.dto.user.OnboardResponse;
+import ccc.keeweapi.dto.user.ProfileMyPageResponse;
+import ccc.keeweapi.dto.user.ProfileUpdateRequest;
+import ccc.keeweapi.dto.user.ProfileUpdateResponse;
+import ccc.keeweapi.dto.user.UnblockUserResponse;
+import ccc.keeweapi.dto.user.UploadProfilePhotoResponse;
 import ccc.keeweapi.utils.SecurityUtil;
 import ccc.keewedomain.dto.user.FollowCheckDto;
 import ccc.keewedomain.persistence.domain.title.TitleAchievement;
@@ -12,8 +25,9 @@ import ccc.keewedomain.persistence.domain.user.Follow;
 import ccc.keewedomain.persistence.domain.user.User;
 import ccc.keewedomain.persistence.repository.utils.CursorPageable;
 import ccc.keewedomain.service.challenge.query.ChallengeParticipateQueryDomainService;
-import ccc.keewedomain.service.user.ProfileDomainService;
 import ccc.keewedomain.service.user.UserDomainService;
+import ccc.keewedomain.service.user.command.ProfileCommandDomainService;
+import ccc.keewedomain.service.user.query.ProfileQueryDomainService;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Set;
@@ -28,27 +42,27 @@ import org.springframework.web.multipart.MultipartFile;
 @Service
 @RequiredArgsConstructor
 public class ProfileApiService {
-
-    private final ProfileDomainService profileDomainService;
     private final UserDomainService userDomainService;
     private final ChallengeParticipateQueryDomainService challengeParticipateQueryDomainService;
     private final ProfileAssembler profileAssembler;
+    private final ProfileQueryDomainService profileQueryDomainService;
+    private final ProfileCommandDomainService profileCommandDomainService;
 
     @Transactional
     public OnboardResponse onboard(OnboardRequest request) {
-        User user = profileDomainService.onboard(profileAssembler.toOnboardDto(request));
+        User user = profileCommandDomainService.onboard(profileAssembler.toOnboardDto(request));
         return profileAssembler.toOnboardResponse(user);
     }
 
     @Transactional
     public FollowToggleResponse toggleFollowership(Long targetId) {
-        boolean following = profileDomainService.toggleFollowership(profileAssembler.toFollowToggleDto(targetId));
+        boolean following = profileCommandDomainService.toggleFollowership(profileAssembler.toFollowToggleDto(targetId));
         return profileAssembler.toFollowToggleResponse(following);
     }
 
     @Transactional
     public UploadProfilePhotoResponse uploadProfilePhoto(MultipartFile imageFile) {
-        String image = profileDomainService.uploadProfilePhoto(profileAssembler.toUploadProfilePhotoDto(imageFile));
+        String image = profileCommandDomainService.uploadProfilePhoto(profileAssembler.toUploadProfilePhotoDto(imageFile));
         return UploadProfilePhotoResponse.of(image);
     }
 
@@ -57,9 +71,9 @@ public class ProfileApiService {
         User targetUser = userDomainService.getUserByIdWithInterests(targetId);
         Long userId = SecurityUtil.getUserId();
 
-        boolean isFollowing = profileDomainService.getFollowingTargetIdSet(FollowCheckDto.of(userId, targetId));
-        Long followerCount = profileDomainService.getFollowerCount(targetUser);
-        Long followingCount = profileDomainService.getFollowingCount(targetUser);
+        boolean isFollowing = profileQueryDomainService.isFollowing(FollowCheckDto.of(userId, targetId));
+        Long followerCount = profileQueryDomainService.getFollowerCount(targetUser);
+        Long followingCount = profileQueryDomainService.getFollowingCount(targetUser);
 
         String challengeName = challengeParticipateQueryDomainService.findCurrentParticipationByUserId(targetId)
                 .map(challengeParticipation -> challengeParticipation.getChallenge().getName())
@@ -70,36 +84,36 @@ public class ProfileApiService {
 
     @Transactional(readOnly = true)
     public MyPageTitleResponse getMyPageTitles(Long userId) {
-        List<TitleAchievement> titleAchievements = profileDomainService.getTitleAchievements(userId, MY_PAGE_TITLE_LIMIT);
-        Long total = profileDomainService.getAchievedTitleCount(userId);
+        List<TitleAchievement> titleAchievements = profileQueryDomainService.getTitleAchievements(userId, MY_PAGE_TITLE_LIMIT);
+        Long total = profileQueryDomainService.getAchievedTitleCount(userId);
 
         return profileAssembler.toMyPageTitleResponse(total, titleAchievements);
     }
 
     @Transactional(readOnly = true)
     public AllAchievedTitleResponse getAllAchievedTitles(Long userId) {
-        List<TitleAchievement> titleAchievements = profileDomainService.getTitleAchievements(userId, Integer.MAX_VALUE);
+        List<TitleAchievement> titleAchievements = profileQueryDomainService.getTitleAchievements(userId, Integer.MAX_VALUE);
 
         return profileAssembler.toAllAchievedTitleResponse(SecurityUtil.getUser(), titleAchievements);
     }
 
     @Transactional(readOnly = true)
     public FollowUserListResponse getFollowers(Long userId, CursorPageable<LocalDateTime> cPage) {
-        List<Follow> follows = profileDomainService.getFollowers(userId, cPage);
+        List<Follow> follows = profileQueryDomainService.getFollowers(userId, cPage);
 
         return getFollowUser(follows, Follow::getFollower);
     }
 
     @Transactional(readOnly = true)
     public FollowUserListResponse getFollowees(Long userId, CursorPageable<LocalDateTime> cPage) {
-        List<Follow> follows = profileDomainService.getFollowees(userId, cPage);
+        List<Follow> follows = profileQueryDomainService.getFollowees(userId, cPage);
 
         return getFollowUser(follows, Follow::getFollowee);
     }
 
     @Transactional
     public BlockUserResponse blockUser(Long blockedUserId) {
-        Long blockUserId = profileDomainService.blockUser(SecurityUtil.getUserId(), blockedUserId);
+        Long blockUserId = profileCommandDomainService.blockUser(SecurityUtil.getUserId(), blockedUserId);
         return profileAssembler.toBlockUserResponse(blockedUserId);
     }
 
@@ -108,30 +122,30 @@ public class ProfileApiService {
                 .map(followUserFunction)
                 .collect(Collectors.toList());
 
-        Set<Long> followingIdSet = profileDomainService.getFollowingTargetIdSet(SecurityUtil.getUser(), users);
+        Set<Long> followingIdSet = profileQueryDomainService.getFollowingTargetIdSet(SecurityUtil.getUser(), users);
         return profileAssembler.toFollowUserListResponse(follows, users, followingIdSet);
     }
 
     @Transactional
     public UnblockUserResponse unblockUser(Long blockedUserId) {
-        Long unblockUserId = profileDomainService.unblockUser(SecurityUtil.getUserId(), blockedUserId);
+        Long unblockUserId = profileCommandDomainService.unblockUser(SecurityUtil.getUserId(), blockedUserId);
         return profileAssembler.toUnblockUserResponse(unblockUserId);
     }
 
     @Transactional(readOnly = true)
     public MyBlockUserListResponse getMyBlockList() {
-        List<Block> blocks = profileDomainService.findBlocksByUserId(SecurityUtil.getUserId());
+        List<Block> blocks = profileQueryDomainService.findBlocksByUserId(SecurityUtil.getUserId());
         return profileAssembler.toMyBlockUserListResponse(blocks);
     }
 
     @Transactional
     public ProfileUpdateResponse updateProfile(ProfileUpdateRequest request) {
-        User user = profileDomainService.updateProfile(SecurityUtil.getUserId(), profileAssembler.toProfileUpdateDto(request));
+        User user = profileCommandDomainService.updateProfile(SecurityUtil.getUserId(), profileAssembler.toProfileUpdateDto(request));
         return profileAssembler.toProfileUpdateResponse(user);
     }
 
     @Transactional(readOnly = true)
     public InterestsResponse getInterests() {
-        return profileAssembler.toInterestsResponse(profileDomainService.getInterests(SecurityUtil.getUser()));
+        return profileAssembler.toInterestsResponse(profileQueryDomainService.getInterests(SecurityUtil.getUser()));
     }
 }
