+++ keewe-domain/src/main/java/ccc/keewedomain/dto/user/FollowToggleDto.java
@@ -1,5 +1,8 @@
 package ccc.keewedomain.dto.user;
 
+import ccc.keewecore.consts.KeeweRtnConsts;
+import ccc.keewecore.exception.KeeweException;
+import java.util.Objects;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -8,4 +11,10 @@ import lombok.Getter;
 public class FollowToggleDto {
     private Long userId;
     private Long targetId;
+
+    public void validateSelfFollowing(FollowToggleDto followDto) {
+        if(Objects.equals(followDto.getTargetId(), followDto.getUserId())) {
+            throw new KeeweException(KeeweRtnConsts.ERR446);
+        }
+    }
 }
